Add Sign-in feature
Front end:

This commit updates the splash page to include a sign in feature.
This feature includes an input for email and password.
Upon submit a POST request is made to the back end.

Back end:
This commit adds the /sign_in route
The POST request is processed and returned.

TO DO:
Use POST data to create user session/sign in ->
Make user sign up form and use data to add user to db.