User Context
This commit adds the Movies.js component that is a WIP.  So far it
it has a click count and contacts random user generator api to get
a users information.

User context has been added. This context will be used to pass down
throughout the component tree to access user details for authentication
and populating data on pages.

TO DO:
Use random generated user to set the user context.